requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
pandas>=2.0.0
pinecone>=2.2.0
